chore: use dropdown component for image selection
The custom select control is nice, but we should use the Dropdown component from
the UI library. This also removes use of some custom colors. Losing the fa icons
is unfortunate, but we don't do this anywhere else in Podman Desktop so I dropped
them as well.

Also noticed a custom a-href, switched to correctly using a Link.

Part of #975.

Signed-off-by: Tim deBoer <git@tdeboer.ca>